use new standards for CEs
