+++ .eslintrc.json
@@ -1,17 +1,42 @@
 {
   "extends": [
     "plugin:github/browser",
-    "plugin:github/es6",
-    "plugin:github/typescript"
+    "plugin:github/recommended",
+    "plugin:github/typescript",
+    "plugin:custom-elements/recommended"
   ],
+  "rules": {
+    "custom-elements/tag-name-matches-class": [
+      "error",
+      {
+        "suffix": "Element"
+      }
+    ],
+    "custom-elements/define-tag-after-class-definition": "off",
+    "custom-elements/no-method-prefixed-with-on": "off",
+    "custom-elements/expose-class-on-global": "off",
+    "import/extensions": ["error", "always"],
+    "import/no-unresolved": "off"
+  },
   "globals": {
     "FilterInputElement": "readable"
   },
   "overrides": [
     {
+      "files": "src/*-define.ts",
+      "rules": {
+        "@typescript-eslint/no-namespace": "off"
+      }
+    },
+    {
       "files": "test/**/*.js",
       "rules": {
-        "github/unescaped-html-literal": "off"
+        "github/unescaped-html-literal": "off",
+        "github/no-inner-html": "off",
+        "i18n-text/no-en": "off"
+      },
+      "env": {
+        "mocha": true
       }
     }
   ]
