--- prettier.config.js
@@ -1,1 +1,0 @@
-module.exports = require('eslint-plugin-github/prettier.config')
